Ensure on_show_section callback for enabled sections (default)
